Support creation of a multisample framebuffer for presentation
WindowTraits now has a samples member, which, if set to something
non-default, will cause a render pass and frame buffer to be created
with a multisampled image. The render pass is set up to resolve that
image. Note that a viewer configured like this will only work with a
scenegraph with the appropriate MultisampleState set up; it is an
error to render with a mismatched sample state.

These changes also expand on using the FrameAssembly class to
encapsulate the supply of render passes and frame buffers to the
RenderGraph class. A VkClearValue object must be supplied for the
multisampled attachment too, so clear values have been added to the
FrameAssembly interface.