Create and install shared library with version on MinGW
In package-oriented distributions like openSUSE, shared libraries are
usually installed with a version to better handle dependencies between
packages.

The main rationale is to allow for two or more (incompatible)
versions of a shared library, such as libfoo.so.1 and libfoo.so.2 to be
independently selectively installable (and trackable by rpm) at the same
time without causing package conflicts.

To achieve this also for MinGW based packages installed on a distribution,
the shared library 'vsg' for this compiler is also created and installed
with version.