[FIXED] invalid recycling of descriptor sets by doing a value compare of descriptorSetLayout
Reusing descriptor sets allocated with a differently defined layout is practically impossible due to various Vulkan requirements. Most importantly, when binding descriptor sets, the descriptor set layout in the pipeline layout must be the same as, or identically defined as, the layout that the bound descriptor set was allocated with (VUID-vkCmdBindDescriptorSets-pDescriptorSets-00358).