Don't create descriptor set bindings with null data
In GraphicsPipelineConfigurator::copyTo(), don't create a
BindDescriptorSet command for descriptor sets that contain null buffer
or image data, and emit a warning if a descriptor set contains a mix
of null values and valid data. The common case with
GraphicsPipelineConfigurator is to bind a pipeline and all the
necessary descriptor sets at the same time (in the same StateGroup);
this commit supports binding some descriptor sets earlier than the
pipeline by not rebinding them with some default value.