Store the ImageInfo for font textures in the font
This is in place of only storing the Data for the font texture and
relying on SharedObjects to find an already created font texture. The
font glyph metrics weren't stored as data at all, but were recreated
on every creation of a GpuLayoutTechnique.

At the moment this change results in a massive memory savings in
complex scenes. There seems to be a bug in SharedObjects in that it
can't pick up the images in a font if they are recreated on each use,
resulting in gigabytes of extra texture use. Even if SharedObjects is
fixed, it is cleaner to store the images in the font.

The glyph metrics texture should probably be a storage buffer
instead.