Make it easier to use separate samplers and textures
Vulkan optionally lets you set up textures and samplers separately, and only combine them in the shader, just like is mandatory in older Direct3D versions (I've not checked D3D12, but I'd assume it's optional just like in Vulkan).

This was already possible with the VSG, but was a pain to set up as there was nothing saying which hoops needed to be jumped through.
This adds a couple of convenience constructors to DescriptorImage that do the hoop-jumping automatically.