Initial spotlight shadow implementation
It doesn't yet attempt to restrict light space to just cover the view frustum or regions of interest, which in turn also means that using more than one shadow map is pointless as they'll all end up the same.
Plenty of engines and middleware frameworks leave it there, but it can be useful to light a scene with a distant spot light instead of a directional light, and then cascaded shadow mapping is just as useful.

There's also no way to set the near plane distance, so it's just a metre from the light, which I chose arbitrarily.
Something needs to be done, as obviously, it can't be zero, and the closer it is to the light, the more depth precision gets thrown away.

Other than that, everything's fine.