Better near plane and bias
We still don't know the 'right' near plane for the shadow map, but using a sensible near/far ratio (in this case, chosen to match the VSG default for the main camera), should give a sensible result more often than using a fixed value of a metre.

Doing this revealed that the existing bias approach was wildly inappropriate for spot lights.
Because of the perspective projection, and reverse-z approach, the fixed light-space offset would be several metres at middling distances from the light, but too small to make any meaningful difference close to the light.
Switching to a scale instead makes the bias small where floats have lots of precision, and larger where they have less, although it doesn't work as well as a logarithmic transform (which is impossible with matrices) would.
In the olden days, this could have been dealt with with polygon offset, as that operates in terms of the smallest resolvable difference of the depth buffer format, but that gets funky with floating-point depth buffers as the smallest resolvable difference is no longer constant across a triangle.