Check and report errors for Win32 functions that may fail
Skips a couple in the Win32_Window destructor. One will throw because of a bug that's on the tracker, so this should be temporary. The other's because there's a relatively benign failure as the API won't let you unregister a window class while there are still windows using it, which is what we want.