Optimise shaders
Unfortunately, on my conveniently-available hardware (AMD running Windows), it doesn't seem to make any difference.
It's definitely changing the bytecode that's sent to the driver (I checked that thoroughly after an earlier revision of this work turned out to only actually be optimising the first shader it was given)

I don't know whether that's because:
* AMD's Windows driver already optimises SPIR-V bytecode at least this well before running it. This is the most plausible explanation to me, though.
* The VSG's stock shader isn't particularly amenable to optimisation.
* I've not tested it on anything where shader execution (except unavoidable bits) takes a meaningful amount of time.

It would make a degree of sense for other people to test this on other machines - I'd be unsurprised if embedded platforms didn't bother optimising the SPIR-V they're given, but I don't have any such hardware myself.