Reorder API dump and validation layers
The validation layers eat and alter some Vulkan calls, so a combined log with both layers will have mismatched information.

In particular, vkSetDebugUtilsObjectNameEXT calls do not make it through the validation layers, so the API dump layer won't be able to see any object names or be able dump the calls that set them, and the validation layer internally remaps object IDs, so the IDs in any validation errors won't match those in the dumped API calls.

Ensuring the calls go through the dump layer before the validation layers ensures all calls are seen and can be dumped, and that the public object IDs that the validation layers expose to the application are the same as end up in the API dump.