Add ref_ptr comparison operators to observer_ptr
Because C++ doesn't like doing implicit conversions when templates are involved, ref_ptr and observer_ptr weren't directly comparable even when holding the same type.

Having seen an app that spent a whole third of its launch time acquiring locks to convert observer_ptrs to ref_ptrs in order to compare them to a ref_ptr, I think this is a footgun and providing a single obvious correct way to deal with that situation with good performance would be a good idea.