Simplify observer_ptr comparison
The comparison of _ptr is redundant as it can only be different if _auxiliary is different, and we don't care if it's the same unless _auxiliary is the same.

Annoyingly, the same simplification can't apply to the other comparison operators.
operator< will give a different order to other kinds of pointer comparison if it only cares about _auxiliary, and when comparing other kinds of pointer, we need to check that rhs isn't null before dereferencing it to get its auxiliary pointer, so still need three comparisons.
We could do a different three comparisons, but this is the easiest combination to read that works.