Add missing &
I'm not entirely sure how these were missed, but there were still ref and unref calls I wasn't expecting on a hot path.
This takes a further five seconds off the startup time of the app I'm working on, so between this and the previous PR, that's nearly half the launch time gone.