Avoid reusing semaphores that may still be in flight
Fixes validation error `VUID-vkQueueSubmit-pSignalSemaphores-00067`

At the moment, this is a little bit too enthusiastic when multiple windows are affected by the same RecordAndSubmitTask.
It would be sufficient to use any one of their `renderFinishedSemaphore`s, but one's used from every window.
This shouldn't make a major difference to anything, as they'll all be signalled at the same time and waited on by the same presentation, but in principle, redundant semaphores might have some overhead.

If this change isn't satisfactory and some things warrant rearranging, then that might become moot anyway, so I've not done anything about it yet and will wait for feedback before doing so.