Do not set RELEASE_RESOURCES flag when resetting command pools by default
This gives an up to 10x performance improvement with AMD's proprietary Windows drivers, although it doesn't seem to make any difference with radv.
I'd imagine that other drivers will see an impact somewhere between the two.

The flag signals to the driver that it should deallocate the memory backing all command buffers from the pool, potentially both in system RAM and VRAM, and potentially means that if command buffers are reused, they must be reallocated from scratch.
This may mean additional CPU-GPU synchronisation, and so it's a heavyweight operation only intended to be used occasionally, e.g. after a large scene is unloaded, not after every frame when it's likely similarly-sized command buffers are about to be used again.