Make intersectors reusable
A lot of time during intersector traversal is spent doing allocations and deallocations.
This can be avoided by just reusing the existing allocations.
The easiest way to do that is to reuse the thing that owns the allocations.

Helpfully, just resizing a vector doesn't throw away its existing allocation, as usually a vector is used for things about the same size as the last time they were used.

This doesn't totally remove all allocations - the array state stack has vectors of matrices in each element, so when elements are removed, those are destroyed and reallocated the next frame, and the intersection vector also has node paths etc. in each element.
Eliminating these would be more challenging.

With the test data I've tried, and using the accelerated IntersectionOptimizedVertexDraw class, reusing the same LineSegmentIntersector instance with this change knocks about 15% off the time to do a lot of intersection queries.
20% of the time is still spent doing allocations and deallocations, though.