Fix compilation as DLL with MSVC
MSVC apparently got confused by the mixed approaches for inheriting constructors, so switched InheritBase to a using declaration as that's the simpler syntax.

This then upset GCC because 'an inherited constructor is not a candidate for initialization from an expression of the same or derived type', so the explicit from-ArrayState constructors required for cloneArrayState stopped being available until they were explicitly reimplemented, and then that meant that DisplacementMapArrayState had to expose it with the copyop.